alternative workaround against #431 - <err> [1000425] pgsql_filecopy_add: INSERT INTO FileCopy (inumber, hostname) VALUES ($1, $2): ERROR:  duplicate key value violates unique constraint "filecopy_pkey"#012DETAIL:  Key (inumber, hostname)=(391680, rgfsd109) already exists.
the workaround in r6591 may have a problem, if a transaction which has such INSERT statement includes other SQL statements (see such symptom about dead_file_copy in comment:4:ticket:407).
this change is a sequel to r6597, and fixes the problem by implementing the workaround in SQL, similar to r6599.


git-svn-id: svn://svn.code.sf.net/p/gfarm/code/gfarm_v2/branches/2.5@6600 4a5e1ece-6a47-475c-b646-731d720b41d0