merge r8994 from the 2.6 branch:
fix a critical problem #754 - sharedsecret configuration allows unauthenticated access in a certain case

this problem was caused by the change for #740 - sharedsecret may not retry to connect to gfmd, when gfmd is restarted,
and found by tatebe@-san.
this happens when getpwnam_r(3) returns 0 with pwd == NULL,
and this cause is found by tatebe@-san, too.


git-svn-id: svn://svn.code.sf.net/p/gfarm/code/gfarm_v2/trunk@8995 4a5e1ece-6a47-475c-b646-731d720b41d0