resolve #806 configure.ac: incorrect host_os_name
the patch in the ticket isn't applied as is, though.
because it brokes netbsd case, for example.
i.e.
both "netbsd6.1.4" and "netbsdelf3.0" should be converted "netbsd",
but the results of the patch are: "netbsd6.1.4" and "netbsdelf3.0"
instead of "netbsd".

the patch is originally written by NIIBE Yutaka,
and #806 is reported by https://sourceforge.net/u/ownerless/.
Thanks!


git-svn-id: svn://svn.code.sf.net/p/gfarm/code/gfarm_v2/trunk@9437 4a5e1ece-6a47-475c-b646-731d720b41d0